fix about 100 warnings
svn path=/trunk/kdeedu/kstars/; revision=283658

+++ kstars/indi/fli/libfli-usb-sys-linux.c
@@ -137,7 +137,7 @@ static long linux_bulktransfer(flidev_t dev, int ep, void *buf, long *len)
   if (ioctl(io->fd, USBDEVFS_RELEASEINTERFACE, &iface))
     return -errno;
 
-  if (*len != bulk.len)
+  if (*len != (long) bulk.len)
     return -errno;
   else
     return 0;
